+++ Nexmo.Api.Test.Integration/ShortCodeTest.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Configuration;
 using NUnit.Framework;
 
 namespace Nexmo.Api.Test.Integration
@@ -12,7 +11,7 @@ namespace Nexmo.Api.Test.Integration
         {
             var response = ShortCode.RequestTwoFactorAuth(new ShortCode.TwoFactorAuthRequest
             {
-                to = ConfigurationManager.AppSettings["test_number"],
+                to = Configuration.Instance.Settings["test_number"],
                 pin = 1467
             });
             Assert.AreEqual("1", response.message_count);
@@ -23,7 +22,7 @@ namespace Nexmo.Api.Test.Integration
         {
             var response = ShortCode.RequestAlert(new ShortCode.AlertRequest
             {
-                to = ConfigurationManager.AppSettings["test_number"],
+                to = Configuration.Instance.Settings["test_number"],
             }, new Dictionary<string, string>
             {
                 {"mcount", "xyz123"}