+++ Nexmo.Api.Test.Integration/SmsTest.cs
@@ -1,5 +1,4 @@
-﻿using System.Configuration;
-using NUnit.Framework;
+﻿using NUnit.Framework;
 
 namespace Nexmo.Api.Test.Integration
 {
@@ -11,11 +10,11 @@ namespace Nexmo.Api.Test.Integration
         {
             var results = SMS.Send(new SMS.SMSRequest
             {
-                from = ConfigurationManager.AppSettings["nexmo_number"],
-                to = ConfigurationManager.AppSettings["test_number"],
+                from = Configuration.Instance.Settings["nexmo_number"],
+                to = Configuration.Instance.Settings["test_number"],
                 text = "this is a test"
             });
-            Assert.AreEqual(ConfigurationManager.AppSettings["test_number"], results.messages[0].to);
+            Assert.AreEqual(Configuration.Instance.Settings["test_number"], results.messages[0].to);
             Assert.AreEqual("0", results.messages[0].status);
         }
     }