+++ Nexmo.Api.Test.Unit/MockedWebTest.cs
@@ -1,29 +1,92 @@
 ﻿using System;
-using System.Configuration;
-using Moq;
-using Nexmo.Api.Request;
+using System.Collections.Generic;
+using System.Linq;
+using System.Net;
+using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
 using NUnit.Framework;
 
 namespace Nexmo.Api.Test.Unit
 {
-    internal class MockedWebTest
+    [SetUpFixture]
+    internal class MockedWebTestSetup
     {
-        protected Mock<IHttpWebRequestFactory> _mock;
-        protected Mock<IHttpWebRequest> _request;
-
-        protected string ApiUrl = ConfigurationManager.AppSettings["Nexmo.Url.Api"];
-        protected string RestUrl = ConfigurationManager.AppSettings["Nexmo.Url.Rest"];
-        protected string ApiKey = ConfigurationManager.AppSettings["Nexmo.api_key"];
-        protected string ApiSecret = ConfigurationManager.AppSettings["Nexmo.api_secret"];
-
-        [SetUp]
+        [OneTimeSetUp]
         public void Setup()
         {
-            _mock = new Mock<IHttpWebRequestFactory>();
-            _request = new Mock<IHttpWebRequest>();
-            _mock.Setup(x => x.CreateHttp(It.IsAny<Uri>()))
-                .Returns<Uri>(r => _request.Object);
-            ApiRequest.WebRequestFactory = _mock.Object;
+            Configuration.Instance.ClientHandler = new FakeClientHandler();
+        }
+    }
+
+    internal class MockedWebTest
+    {
+        protected string ApiUrl = Configuration.Instance.Settings["Nexmo.Url.Api"];
+        protected string RestUrl = Configuration.Instance.Settings["Nexmo.Url.Rest"];
+        protected string ApiKey = Configuration.Instance.Settings["Nexmo.api_key"];
+        protected string ApiSecret = Configuration.Instance.Settings["Nexmo.api_secret"];
+
+        protected void SetExpect(string uri, string response, string content = "")
+        {
+            var fakeClientHandler = (FakeClientHandler)Configuration.Instance.ClientHandler;
+            fakeClientHandler.ExpectedUris.Push(uri);
+            fakeClientHandler.ExpectedResponses.Push(response);
+            fakeClientHandler.ExpectedContent.Push(content);
+        }
+
+        protected void SetExpectStatus(HttpStatusCode code)
+        {
+            var fakeClientHandler = (FakeClientHandler)Configuration.Instance.ClientHandler;
+            fakeClientHandler.ExpectedStatusCodes.Push(code);
+        }
+    }
+
+    internal class FakeClientHandler : HttpClientHandler
+    {
+        internal Stack<string> ExpectedUris { get; } = new Stack<string>();
+        internal Stack<string> ExpectedContent { get; } = new Stack<string>();
+        internal Stack<HttpStatusCode> ExpectedStatusCodes { get; } = new Stack<HttpStatusCode>();
+        internal Stack<string> ExpectedResponses { get; } = new Stack<string>();
+
+        protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
+        {
+            var uri = ExpectedUris.Pop();
+            var responseContent = ExpectedResponses.Pop();
+            var requestContent = string.Empty;
+            if (ExpectedContent.Any())
+            {
+                requestContent = ExpectedContent.Pop();
+            }
+
+            if (!string.Equals(request.RequestUri.AbsoluteUri, uri, StringComparison.OrdinalIgnoreCase))
+                return AssertFail();
+
+            if (!string.IsNullOrEmpty(requestContent))
+            {
+                var contentReadTask = request.Content.ReadAsStringAsync();
+                contentReadTask.Wait(cancellationToken);
+                if (!string.Equals(contentReadTask.Result, requestContent, StringComparison.OrdinalIgnoreCase))
+                {
+                    return AssertFail();
+                }
+            }
+
+            var responseCode = HttpStatusCode.OK;
+            if (ExpectedStatusCodes.Any())
+            {
+                responseCode = ExpectedStatusCodes.Pop();
+            }
+            var response = new HttpResponseMessage(responseCode)
+            {
+                Content = new StringContent(responseContent)
+            };
+            return Task.FromResult(response);
+        }
+
+        private static Task<HttpResponseMessage> AssertFail()
+        {
+            var response = new HttpResponseMessage(HttpStatusCode.NotFound);
+            return Task.FromResult(response);
         }
     }
 }