+++ Nexmo.Api.Test.Unit/NumberTest.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.IO;
-using System.Linq;
-using System.Text;
-using Moq;
-using Nexmo.Api.Request;
+﻿using System.Linq;
 using NUnit.Framework;
 
 namespace Nexmo.Api.Test.Unit
@@ -14,17 +9,14 @@ namespace Nexmo.Api.Test.Unit
         [Test]
         public void should_search_numbers()
         {
-            var resp = new Mock<IWebResponse>();
-            resp.Setup(e => e.GetResponseStream()).Returns(new MemoryStream(Encoding.UTF8.GetBytes("{\"count\":177,\"numbers\":[{\"country\":\"US\",\"msisdn\":\"15102694548\",\"type\":\"mobile-lvn\",\"features\":[\"SMS\",\"VOICE\"],\"cost\":\"0.67\"},{\"country\":\"US\",\"msisdn\":\"17088568490\",\"type\":\"mobile-lvn\",\"features\":[\"SMS\",\"VOICE\"],\"cost\":\"0.67\"},{\"country\":\"US\",\"msisdn\":\"17088568491\",\"type\":\"mobile-lvn\",\"features\":[\"SMS\",\"VOICE\"],\"cost\":\"0.67\"},{\"country\":\"US\",\"msisdn\":\"17088568492\",\"type\":\"mobile-lvn\",\"features\":[\"SMS\",\"VOICE\"],\"cost\":\"0.67\"},{\"country\":\"US\",\"msisdn\":\"17088568973\",\"type\":\"mobile-lvn\",\"features\":[\"SMS\",\"VOICE\"],\"cost\":\"0.67\"}]}")));
-            _request.Setup(e => e.GetResponse()).Returns(resp.Object);
+            SetExpect($"{RestUrl}/number/search/?country=US&api_key={ApiKey}&api_secret={ApiSecret}&",
+"{\"count\":177,\"numbers\":[{\"country\":\"US\",\"msisdn\":\"15102694548\",\"type\":\"mobile-lvn\",\"features\":[\"SMS\",\"VOICE\"],\"cost\":\"0.67\"},{\"country\":\"US\",\"msisdn\":\"17088568490\",\"type\":\"mobile-lvn\",\"features\":[\"SMS\",\"VOICE\"],\"cost\":\"0.67\"},{\"country\":\"US\",\"msisdn\":\"17088568491\",\"type\":\"mobile-lvn\",\"features\":[\"SMS\",\"VOICE\"],\"cost\":\"0.67\"},{\"country\":\"US\",\"msisdn\":\"17088568492\",\"type\":\"mobile-lvn\",\"features\":[\"SMS\",\"VOICE\"],\"cost\":\"0.67\"},{\"country\":\"US\",\"msisdn\":\"17088568973\",\"type\":\"mobile-lvn\",\"features\":[\"SMS\",\"VOICE\"],\"cost\":\"0.67\"}]}");
+
             var results = Number.Search(new Number.SearchRequest
             {
                 country = "US"
             });
 
-            _mock.Verify(h => h.CreateHttp(new Uri(
-                string.Format("{0}/number/search/?country=US&api_key={1}&api_secret={2}&", RestUrl, ApiKey, ApiSecret))),
-                Times.Once);
             Assert.AreEqual(177, results.count);
             Assert.AreEqual(5, results.numbers.Count());
         }
@@ -32,20 +24,11 @@ namespace Nexmo.Api.Test.Unit
         [Test]
         public void should_buy_number()
         {
-            var resp = new Mock<IWebResponse>();
-            resp.Setup(e => e.GetResponseStream()).Returns(new MemoryStream(Encoding.UTF8.GetBytes("{\"error-code\":\"200\",\"error-code-label\":\"success\"}")));
-            var postDataStream = new MemoryStream();
-            _request.Setup(e => e.GetRequestStream()).Returns(postDataStream);
-            _request.Setup(e => e.GetResponse()).Returns(resp.Object);
-            var result = Number.Buy("US", "17775551212");
+            SetExpect($"{RestUrl}/number/buy",
+"{\"error-code\":\"200\",\"error-code-label\":\"success\"}",
+$"country=US&msisdn=17775551212&api_key={ApiKey}&api_secret={ApiSecret}&");
 
-            _mock.Verify(h => h.CreateHttp(new Uri(
-                string.Format("{0}/number/buy", RestUrl))),
-                Times.Once);
-            postDataStream.Position = 0;
-            var sr = new StreamReader(postDataStream);
-            var postData = sr.ReadToEnd();
-            Assert.AreEqual(string.Format("country=US&msisdn=17775551212&api_key={0}&api_secret={1}&", ApiKey, ApiSecret), postData);
+            var result = Number.Buy("US", "17775551212");
 
             Assert.AreEqual("200", result.ErrorCode);
         }
@@ -53,11 +36,10 @@ namespace Nexmo.Api.Test.Unit
         [Test]
         public void should_update_number()
         {
-            var resp = new Mock<IWebResponse>();
-            resp.Setup(e => e.GetResponseStream()).Returns(new MemoryStream(Encoding.UTF8.GetBytes("{\"error-code\":\"200\",\"error-code-label\":\"success\"}")));
-            var postDataStream = new MemoryStream();
-            _request.Setup(e => e.GetRequestStream()).Returns(postDataStream);
-            _request.Setup(e => e.GetResponse()).Returns(resp.Object);
+            SetExpect($"{RestUrl}/number/update",
+"{\"error-code\":\"200\",\"error-code-label\":\"success\"}",
+$"country=US&msisdn=17775551212&moHttpUrl=https%3a%2f%2ftest.test.com%2fmo&moSmppSysType=inbound&api_key={ApiKey}&api_secret={ApiSecret}&");
+
             var result = Number.Update(new Number.NumberUpdateCommand
             {
                 country = "US",
@@ -66,34 +48,17 @@ namespace Nexmo.Api.Test.Unit
                 moSmppSysType = "inbound"
             });
 
-            _mock.Verify(h => h.CreateHttp(new Uri(
-                string.Format("{0}/number/update", RestUrl))),
-                Times.Once);
-            postDataStream.Position = 0;
-            var sr = new StreamReader(postDataStream);
-            var postData = sr.ReadToEnd();
-            Assert.AreEqual(string.Format("country=US&msisdn=17775551212&moHttpUrl=https%3a%2f%2ftest.test.com%2fmo&moSmppSysType=inbound&api_key={0}&api_secret={1}&", ApiKey, ApiSecret), postData);
-
             Assert.AreEqual("200", result.ErrorCode);
         }
 
         [Test]
         public void should_cancel_number()
         {
-            var resp = new Mock<IWebResponse>();
-            resp.Setup(e => e.GetResponseStream()).Returns(new MemoryStream(Encoding.UTF8.GetBytes("{\"error-code\":\"200\",\"error-code-label\":\"success\"}")));
-            var postDataStream = new MemoryStream();
-            _request.Setup(e => e.GetRequestStream()).Returns(postDataStream);
-            _request.Setup(e => e.GetResponse()).Returns(resp.Object);
-            var result = Number.Cancel("US", "17775551212");
+            SetExpect($"{RestUrl}/number/cancel",
+"{\"error-code\":\"200\",\"error-code-label\":\"success\"}",
+$"country=US&msisdn=17775551212&api_key={ApiKey}&api_secret={ApiSecret}&");
 
-            _mock.Verify(h => h.CreateHttp(new Uri(
-                string.Format("{0}/number/cancel", RestUrl))),
-                Times.Once);
-            postDataStream.Position = 0;
-            var sr = new StreamReader(postDataStream);
-            var postData = sr.ReadToEnd();
-            Assert.AreEqual(string.Format("country=US&msisdn=17775551212&api_key={0}&api_secret={1}&", ApiKey, ApiSecret), postData);
+            var result = Number.Cancel("US", "17775551212");
 
             Assert.AreEqual("200", result.ErrorCode);
         }