+++ Nexmo.Api/PemParse.cs
@@ -51,6 +51,7 @@ THE SOFTWARE.
 
 using System;
 using System.IO;
+using System.Runtime.InteropServices;
 using System.Security.Cryptography;
 using System.Text;
 
@@ -61,7 +62,7 @@ namespace Nexmo.Api
         private const string pemprivheader = "-----BEGIN RSA PRIVATE KEY-----";
         private const string pemprivfooter = "-----END RSA PRIVATE KEY-----";
 
-        public static RSACryptoServiceProvider DecodePEMKey(string pemstr)
+        public static RSA DecodePEMKey(string pemstr)
         {
             if (!pemstr.StartsWith(pemprivheader) || !pemstr.EndsWith(pemprivfooter)) return null;
             var pemprivatekey = DecodeOpenSSLPrivateKey(pemstr);
@@ -145,93 +146,120 @@ namespace Nexmo.Api
             //////////}
         }
 
-        public static RSACryptoServiceProvider DecodeRSAPrivateKey(byte[] privkey)
+        public static RSA DecodeRSAPrivateKey(byte[] privkey)
         {
             byte[] MODULUS, E, D, P, Q, DP, DQ, IQ;
 
             // ---------  Set up stream to decode the asn.1 encoded RSA private key  ------
-            var mem = new MemoryStream(privkey);
-            var binr = new BinaryReader(mem); //wrap Memory Stream with BinaryReader for easy reading
-            byte bt = 0;
-            ushort twobytes = 0;
-            var elems = 0;
-            try
+            using (var mem = new MemoryStream(privkey))
+            using (var binr = new BinaryReader(mem)) //wrap Memory Stream with BinaryReader for easy reading
             {
-                twobytes = binr.ReadUInt16();
-                if (twobytes == 0x8130) //data read as little endian order (actual data order for Sequence is 30 81)
-                    binr.ReadByte(); //advance 1 byte
-                else if (twobytes == 0x8230)
-                    binr.ReadInt16(); //advance 2 bytes
-                else
+                byte bt = 0;
+                ushort twobytes = 0;
+                var elems = 0;
+                try
+                {
+                    twobytes = binr.ReadUInt16();
+                    if (twobytes == 0x8130) //data read as little endian order (actual data order for Sequence is 30 81)
+                        binr.ReadByte(); //advance 1 byte
+                    else if (twobytes == 0x8230)
+                        binr.ReadInt16(); //advance 2 bytes
+                    else
+                        return null;
+
+                    twobytes = binr.ReadUInt16();
+                    if (twobytes != 0x0102) //version number
+                        return null;
+                    bt = binr.ReadByte();
+                    if (bt != 0x00)
+                        return null;
+
+
+                    //------  all private key components are Integer sequences ----
+                    elems = GetIntegerSize(binr);
+                    MODULUS = binr.ReadBytes(elems);
+
+                    elems = GetIntegerSize(binr);
+                    E = binr.ReadBytes(elems);
+
+                    elems = GetIntegerSize(binr);
+                    D = binr.ReadBytes(elems);
+
+                    elems = GetIntegerSize(binr);
+                    P = binr.ReadBytes(elems);
+
+                    elems = GetIntegerSize(binr);
+                    Q = binr.ReadBytes(elems);
+
+                    elems = GetIntegerSize(binr);
+                    DP = binr.ReadBytes(elems);
+
+                    elems = GetIntegerSize(binr);
+                    DQ = binr.ReadBytes(elems);
+
+                    elems = GetIntegerSize(binr);
+                    IQ = binr.ReadBytes(elems);
+
+                    //Console.WriteLine("showing components ..");
+                    //if (verbose)
+                    //{
+                    //    showBytes("\nModulus", MODULUS);
+                    //    showBytes("\nExponent", E);
+                    //    showBytes("\nD", D);
+                    //    showBytes("\nP", P);
+                    //    showBytes("\nQ", Q);
+                    //    showBytes("\nDP", DP);
+                    //    showBytes("\nDQ", DQ);
+                    //    showBytes("\nIQ", IQ);
+                    //}
+
+                    // ------- create RSACryptoServiceProvider instance and initialize with public key -----
+#if NETSTANDARD1_6
+                    RSA RSA;
+                    if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux) ||
+                        RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
+                    {
+                        RSA = new RSAOpenSsl();
+                    }
+                    else
+                    {
+                        RSA = new RSACng();
+                    }
+                    var RSAparams = new RSAParameters
+                    {
+                        Modulus = MODULUS,
+                        Exponent = E,
+                        D = D,
+                        P = P,
+                        Q = Q,
+                        DP = DP,
+                        DQ = DQ,
+                        InverseQ = IQ
+                    };
+                    RSA.ImportParameters(RSAparams);
+                    return RSA;
+#else
+                    var RSA = new RSACryptoServiceProvider();
+                    var RSAparams = new RSAParameters
+                    {
+                        Modulus = MODULUS,
+                        Exponent = E,
+                        D = D,
+                        P = P,
+                        Q = Q,
+                        DP = DP,
+                        DQ = DQ,
+                        InverseQ = IQ
+                    };
+                    RSA.ImportParameters(RSAparams);
+                    return RSA;
+#endif
+                }
+                catch (Exception)
+                {
+                    // TODO: log this!
                     return null;
-
-                twobytes = binr.ReadUInt16();
-                if (twobytes != 0x0102) //version number
-                    return null;
-                bt = binr.ReadByte();
-                if (bt != 0x00)
-                    return null;
-
-
-                //------  all private key components are Integer sequences ----
-                elems = GetIntegerSize(binr);
-                MODULUS = binr.ReadBytes(elems);
-
-                elems = GetIntegerSize(binr);
-                E = binr.ReadBytes(elems);
-
-                elems = GetIntegerSize(binr);
-                D = binr.ReadBytes(elems);
-
-                elems = GetIntegerSize(binr);
-                P = binr.ReadBytes(elems);
-
-                elems = GetIntegerSize(binr);
-                Q = binr.ReadBytes(elems);
-
-                elems = GetIntegerSize(binr);
-                DP = binr.ReadBytes(elems);
-
-                elems = GetIntegerSize(binr);
-                DQ = binr.ReadBytes(elems);
-
-                elems = GetIntegerSize(binr);
-                IQ = binr.ReadBytes(elems);
-
-                //Console.WriteLine("showing components ..");
-                //if (verbose)
-                //{
-                //    showBytes("\nModulus", MODULUS);
-                //    showBytes("\nExponent", E);
-                //    showBytes("\nD", D);
-                //    showBytes("\nP", P);
-                //    showBytes("\nQ", Q);
-                //    showBytes("\nDP", DP);
-                //    showBytes("\nDQ", DQ);
-                //    showBytes("\nIQ", IQ);
-                //}
-
-                // ------- create RSACryptoServiceProvider instance and initialize with public key -----
-                var RSA = new RSACryptoServiceProvider();
-                var RSAparams = new RSAParameters();
-                RSAparams.Modulus = MODULUS;
-                RSAparams.Exponent = E;
-                RSAparams.D = D;
-                RSAparams.P = P;
-                RSAparams.Q = Q;
-                RSAparams.DP = DP;
-                RSAparams.DQ = DQ;
-                RSAparams.InverseQ = IQ;
-                RSA.ImportParameters(RSAparams);
-                return RSA;
-            }
-            catch (Exception)
-            {
-                return null;
-            }
-            finally
-            {
-                binr.Close();
+                }
             }
         }