+++ Nexmo.Api/Request/IHttpWebRequestFactory.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace Nexmo.Api.Request
-{
-    public interface IHttpWebRequestFactory
-    {
-        IHttpWebRequest CreateHttp(Uri uri);
-    }
-}