+++ CHANGELOG.md
@@ -1,4 +1,27 @@
-<a name="1.0.0"></a>
+# 2.0.0 (2016-10-24)
+
+* Dependency marking for netstandard1.6
+
+# 2.0.0-rc2 (2016-10-22)
+
+* Fix JWT generation (key import fail) on OSX/Linux
+
+# 2.0.0-rc1 (2016-10-16)
+
+* NumberInsight basic + standard support
+* NumberVerify control call
+* JWT token generation
+* Application API support
+* Application-based call API support
+* .NET Standard 1.6 support
+* __[BREAKING]__ Moved configuration from app.config to settings.json
+* __[BREAKING]__ Nexmo.Api.Voice static class has been deprecated - you must move to the new Voice calls inside the new Nexmo.Api.Voice namespace. See [the Nexmo docs](https://docs.nexmo.com/voice/voice-api) for details.
+
+JWT notes:
+
+* When registering a new application, make sure you save the private key. This library does not (currently) take care of this for you.
+* Make sure your saved private key is ASCII (not UTF-8, no BOM) - http://stackoverflow.com/questions/1068650/using-awk-to-remove-the-byte-order-mark
+
 # 1.0.0 (2016-03-19)
 
 * Initial release with nuget package